test(auv-cli-invoke): move capture API coverage
